requirements.txt
numpy >= 1.16.4
pandas >= 0.24.2
pillow >= 6.1.0